% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{assay-class}
\alias{assay-class}
\alias{assay}
\title{Assay Class}
\description{
This class defines an assay object
}
\section{Slots}{

\describe{
\item{\code{data}}{list of datasets}

\item{\code{analysis}}{list of analyses}

\item{\code{calibration}}{calibration status (pre- or post-calibration)}

\item{\code{plots}}{list of plots generated}

\item{\code{variables}}{variables available for analysis in dataset}

\item{\code{unique.features}}{list of features provided in dataset}

\item{\code{N}}{list of number of available variables for each feature}

\item{\code{description}}{description}
}}

