% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{calibrate.data}
\alias{calibrate.data}
\title{Calibrate Data}
\usage{
calibrate.data(object, which.assay = NULL, verbose = T)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to calibate.}

\item{verbose}{Logical specifying whether progress is reported.}
}
\value{
Calibration Object
}
\description{
Calibrate data set using fit calibration curves. Calibration curves must exist, having been generated by fit.calibration function.
}
\seealso{
fit.calibration
}
