% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{calibrateData}
\alias{calibrateData}
\title{Calibrate Data}
\usage{
calibrateData(object, which.assay = NULL, verbose = T)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to calibate.}

\item{verbose}{Logical specifying whether progress is reported.}
}
\value{
Calibration Object
}
\description{
Calibrate data set using fit calibration curves. Calibration curves must exist, having been generated by fit.calibration function.
}
\details{
New column is added to calibrated dataset, name cal.phantom, specifying which phantom was used for calibraiton. That is, if the uncalibrated data contains datasets from multiple phantom-types, the 'phantom' column specifies which phantom the values originate from and the 'cal.phantom' column specifies which phantom was used for calibration.
}
\seealso{
\code{\link{fitCalibration}}
}
