% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\docType{class}
\name{calibration-class}
\alias{calibration-class}
\alias{calibration}
\title{Calibration Class}
\description{
This class defines a calibration object
}
\section{Slots}{

\describe{
\item{\code{assays}}{list of assay objects}

\item{\code{metadata}}{input dataset (data.frame)}

\item{\code{current.assay}}{species current assay on which analyses are called}

\item{\code{project.name}}{project name}

\item{\code{misc}}{miscilaneous features}
}}

