% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{clone.assay}
\alias{clone.assay}
\title{Clone assay object}
\usage{
clone.assay(
  object,
  which.assay = NULL,
  cloned.assay.name = NULL,
  set.clone.as.default = FALSE
)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies which assay to clone (character)}

\item{cloned.assay.name}{name of clone. If not specified, "-copy" is appended to name of cloned assay.}

\item{set.clone.as.default}{logical specifying whether to set clone as default assay}
}
\value{
calibration object
}
\description{
Create duplicate of assay within calibration object
}
