% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{cloneAssay}
\alias{cloneAssay}
\title{Clone assay object}
\usage{
cloneAssay(
  object,
  which.assay = NULL,
  cloned.assay.name = NULL,
  set.clone.as.default = FALSE,
  verbose = T
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Specifies which assay to clone (character)}

\item{cloned.assay.name}{Name of clone. If not specified, "-copy" is appended to name of cloned Assay.}

\item{set.clone.as.default}{Logical specifying whether to set clone as default assay}

\item{verbose}{Logical to report progress.}
}
\value{
calibration object
}
\description{
Create duplicate of assay within Calibration Object.
}
