% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{compareCalibrationPlot}
\alias{compareCalibrationPlot}
\title{Compare Calibration Equations}
\usage{
compareCalibrationPlot(
  object,
  which.assay.1,
  which.assay.2,
  which.parameter = NULL,
  which.plot = "scatter",
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  box.alpha = 0.5,
  point.size = 2,
  point.alpha = 1,
  slope.scale = NULL,
  intercept.scale = NULL,
  show.reference.line = T
)
}
\arguments{
\item{object}{Calibration Object.}

\item{which.assay.1}{First Assay for calibraiton comparison. Assumed to have fitted calibration equations.}

\item{which.assay.2}{Second Assay for calibraiton comparison. Assumed to have fitted calibration equations.}

\item{which.parameter}{Character specifying which parameters to plot. If unspecified, one parameter is expected in the dataset.}

\item{which.plot}{Character specifying which plot to generate. One of:
\itemize{
\item scatter - Default. Pairwise scatter plot between slopes and intercepts in both assays. If scatter, spearman rho is calculated and reported.
\item box - Boxplot of slopes and intercepts for each assay. If box,  Kruskal-Wallis rank sum test is conducted and p-value is reported.
}}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins. Default is 0.}

\item{color.end}{Hue in [0,1] at which the viridis color map ends. Default is 1.}

\item{box.alpha}{Numeric [0,1] specifying degree of transparency for box plots. Default is 0.5.}

\item{point.size}{Numeric specifying size of scatter plot points. Default is 2.}

\item{point.alpha}{Numeric [0,1] specifying degree of transparency for points. Default is 1.}

\item{slope.scale}{Range of values [a,b] specifying min (a) and max(a) limits of slope-related plots. Specify as vector c(a,b).}

\item{intercept.scale}{Range of values [a,b] specifying min (a) and max(a) limits of intercept-related plots. Specify as vector c(a,b).}

\item{show.reference.line}{Logical indicating whether reference x=y line is shown as dashed line. Only specified if which.plot is scatter. Default is true.}
}
\value{
ggplot object
}
\description{
Generates boxplot summarizing results generated during svp.analysis. Additionally, Kruskal-Wallis rank sum test is performed (using group.by variable for data stratification) and p-values are shown for each comparison group.
}
\seealso{
\code{\link{calibrationPlot}}, \code{\link{fitCalibration}}
}
