% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{consistency.analysis}
\alias{consistency.analysis}
\title{Evaluate site consistency}
\usage{
consistency.analysis(object, which.assay = NULL, which.data = "uncalibrated")
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}
}
\value{
data.frame
}
\description{
Sites are ranked according to mean-squared error calculated between site-specific parameters and overall median parameter values (pooled across all timePoints and sites).
}
\seealso{
\code{\link{identify.reference}}, \code{\link{consistency.plot}}
}
