% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{consistency.plot}
\alias{consistency.plot}
\title{Visualize site rankings according to mean-squared errors}
\usage{
consistency.plot(
  object,
  which.assay = NULL,
  which.data = "uncalibrated",
  which.plot = "box",
  show.tile.value = F
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}

\item{which.plot}{One of:
\itemize{
\item "box" - boxplot of site-specific parameter rankings
\item "tile" - Default. heatmap of site-specific parameter rankings
}}

\item{show.tile.value}{Logical specifying whether to show site rankings, overlaid on heamap (if which.plot == "tile")}
}
\value{
Character
}
\description{
Site rankings are visualied
}
\seealso{
\code{\link{identify.reference}}, \code{\link{consistency.analysis}}
}
