% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{consistencyAnalysis}
\alias{consistencyAnalysis}
\title{Evaluate site consistency}
\usage{
consistencyAnalysis(object, which.assay = NULL, which.data = "uncalibrated")
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}
}
\value{
data frame
}
\description{
Sites are ranked according to mean-squared error calculated between site-specific parameters and overall median parameter values (pooled across all timePoints and sites).
}
\seealso{
\code{\link{identifyReference}}, \code{\link{consistencyPlot}}
}
