% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{consistencyPlot}
\alias{consistencyPlot}
\title{Visualize site rankings according to mean-squared errors}
\usage{
consistencyPlot(
  object,
  var2plot = "rank",
  which.phantom = NULL,
  which.parameter = NULL,
  group.by = NULL,
  which.assay = NULL,
  which.data = "uncalibrated",
  which.plot = "box",
  color.option = "cividis",
  show.tile.value = F
)
}
\arguments{
\item{object}{Calibration Object}

\item{var2plot}{Character indicating which ranking meausre to plot. On of:
\itemize{
\item "rank" - Default
\item "mse" - Normalized mean-squared errors (mse) are plotted.
}}

\item{which.phantom}{Specify which phantom to plot rankings for. If unspecified and multiple phantoms exist in dataset, function will return error requesting that specific phantom be specified.}

\item{which.parameter}{character indicating which parameters to include. If unspecified, all are included.}

\item{group.by}{Grouping variable}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}

\item{which.plot}{One of:
\itemize{
\item "box" - boxplot of site-specific parameter rankings
\item "tile" - Default. heatmap of site-specific parameter rankings
}}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{show.tile.value}{Logical specifying whether to show site rankings, overlaid on heamap (if which.plot == "tile")}
}
\value{
Character
}
\description{
Site rankings are visualized
}
\seealso{
\code{\link{identifyReference}}, \code{\link{consistencyAnalysis}}
}
