% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCalibrationObject.R
\name{createCalibrationObject}
\alias{createCalibrationObject}
\title{Create Calibration Object}
\usage{
createCalibrationObject(df, assay.name = NULL, project.name = NULL)
}
\arguments{
\item{df}{HR-pQCT dataset (data.frame) containing value column along with descriptors site, section, timePoint, phantom, parameter and/or scanDate}

\item{assay.name}{assay name}

\item{project.name}{project name}
}
\value{
Calibration Object
}
\description{
The Calibration object is the engine that drives each calibration and precision analysis.
It stores all information associated with the dataset, including the original input,
transformations, analyses, results and plots. All that is needed to construct a Calibration
object is raw data formatted as a data.frame. See documentation for details
}
