% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_functions.R
\name{deleteAnalysis}
\alias{deleteAnalysis}
\title{Delete Analysis}
\usage{
deleteAnalysis(object, which.assay, which.analysis, verbose = TRUE)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies assay}

\item{which.analysis}{specifies analysis}

\item{verbose}{print progress}
}
\value{
character
}
\description{
Delete analysis from specified assay of calibration object
}
