% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_functions.R
\name{deleteAssay}
\alias{deleteAssay}
\title{Delete Assay}
\usage{
deleteAssay(object, which.assay, verbose = TRUE)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{which assay to delete (character)}

\item{verbose}{print progress (Deafult = TRUE)}
}
\value{
calibration object with specified assay deleted. If specified assay was current assay, new current assay is assigned.
}
\description{
Delete assay from calibration object
}
