% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{HR-pQCt phantom measure calibration and reproducibility dataset}
\format{A data frame with columns:
\describe{
\item{site}{site/scanner used to obtain meaures (e.g., Montreal, Toronto, etc.)}
\item{timePoint}{time of scan; numeric (e.g., 0, 6, 12)}
\item{section}{phantom section; numeric (e.g., value 1-4)}
\item{parameter}{measured parameter (e.g., Tb. vBMD, Ct. Ar., etc)}
\item{value}{measured value; numeric}
\item{phantom}{phantom type (e.g., EFP, QC1)}
\item{scanDate}{date of scan (e.g., 2018-02-27)}
}}
\source{
Shriners Hospital for Children, Mereo BioPharma
}
\usage{
df
}
\description{
HR-pQCt phantom measure calibration and reproducibility dataset
}
\examples{
\dontrun{
# load data into global enviroment
 df <- df
}
}
\keyword{datasets}
