% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{diagnostic.plot}
\alias{diagnostic.plot}
\title{Diagnostic plot}
\usage{
diagnostic.plot(object, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{which assay to plot}
}
\value{
calibration object
}
\description{
Generates diagnostic curves for evaluation of calibration. Require that data have been calibrated.
}
