% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{diagnosticPlot}
\alias{diagnosticPlot}
\title{Diagnostic plot}
\usage{
diagnosticPlot(
  object,
  which.assay = NULL,
  which.parameter = NULL,
  which.plot = "bar",
  highlight.site = NULL,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  bar.errors = F,
  fix.axis = T,
  return.plt.handle = F
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to use.}

\item{which.parameter}{Character indicating which parameters to plot diagnostics for. If unspecified, all parameters are plotted.}

\item{which.plot}{Character indicating which diagnostic plot to produce. One of:
\itemize{
\item "line" - Line plot used to compare pre and post calibrated values.
\item "bar" - Horizontal barplots used to compare pre and post calibrated values
\item "residual" - Histogram used to visualize pre- and post-calibration residuals (i.e., difference between overall mean and replicate mean)
}}

\item{highlight.site}{Character specifying site to highlight in diagnostic plot (valid for line plots)}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins}

\item{color.end}{Hue in [0,1] at which the viridis color map ends}

\item{bar.errors}{Logical indicating whether to show error bars for barplot (valid for bar plots)}

\item{fix.axis}{Logical indicating whether to fix x-axis. Default is True. (Valid for bar and residual plots)}

\item{return.plt.handle}{Logical specifying whether to return list of plot handles. If TRUE, list of plot handles is returned. If FALSE, plots are printed without returning handle.}
}
\value{
plot handle (see return.plt.handle argument)
}
\description{
Generates diagnostic curves for evaluation of calibration. Require that data have been calibrated.
}
\seealso{
\code{\link{fitCalibration}}, \code{\link{calibrateData}}
}
