% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data_functions.R
\name{filterData}
\alias{filterData}
\title{Filter Dataset}
\usage{
filterData(df, analyze.these)
}
\arguments{
\item{df}{HR-pQCT dataset (data.frame) containing value column along with descriptors site, section, timePoint, phantom, parameter and/or scanDate}

\item{analyze.thesis}{specifies which features to include in analysis}
}
\value{
data.frame
}
\description{
Filter HR-pQCT data to only include features of interest
}
