% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{fit.calibration}
\alias{fit.calibration}
\title{Fit calibration curves}
\usage{
fit.calibration(
  object,
  reference.site = NULL,
  reference.time = "baseline",
  sig.intercept.only = F,
  which.assay = NULL,
  n.signif = 3,
  verbose = T
)
}
\arguments{
\item{object}{Calibration Object}

\item{reference.site}{Character specifying reference site (i.e., reference instrument to calibrated other instruments)}

\item{reference.time}{Character specifying reference time. Does not need to be specified if only one timePoint is available. One of:
\itemize{
\item baseline - Default. All calibration curves are generated using reference site measurements at baseline as the reference timpoint.
\item match - Calibration curves are constructed using matched time points. E.g., Site measurements are baseline are calibrated using reference site measurements at baseline, 6 month measurements are calibrated using references measurements at 6 months, etc...
}}

\item{sig.intercept.only}{Logical specifying whether to include intercept in fitted regression equation.
\itemize{
\item TRUE - Only include intercept in model if significant (p < 0.05).
\item FALSE - Default. Include intercept in model regradless of significance.
}}

\item{which.assay}{Character specifying which assay to fit calibration curves for.}

\item{n.signif}{Number of significant figures to report.}

\item{verbose}{Logical specify whether to report progress.}
}
\value{
Calibration Object
}
\description{
Fit pairwise linear regression between calibration and reference sites, using uncalibrated dataset in specified Assay.
}
