% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.analyses}
\alias{get.analyses}
\alias{get.datasets}
\title{Get dataset names}
\usage{
get.datasets(object, which.assay = NULL)

get.analyses(object, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies which assay to retrieve feature counts from (character)}
}
\value{
character

character
}
\description{
Get names of datasets stored in specified assay in calibration object

Get names of avaialble analyses for specified assay in calibration object
}
