% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.datasets}
\alias{get.datasets}
\title{Get dataset names}
\usage{
get.datasets(object, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies which assay to retrieve feature counts from (character)}
}
\value{
character
}
\description{
Get names of datasets stored in specified assay in calibration object
}
