% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.df.from.list}
\alias{get.df.from.list}
\title{Convert multi-tiered list of data.frames to single-tired list of data.frames}
\usage{
get.df.from.list(a)
}
\arguments{
\item{a}{list of data.frames}
}
\value{
list
}
\description{
conversion to single-tiered list of data.frames only supported for 3 tiers currently.
}
