% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.features}
\alias{get.features}
\title{Get Features}
\usage{
get.features(object, verbose = TRUE, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{verbose}{print progress}

\item{which.assay}{specifies which assay to retrieve features from (character)}
}
\value{
list of unique features
}
\description{
Get list of unique features from specified assay in calibration object
}
