% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{get.results.table}
\alias{get.results.table}
\title{Get Results Table}
\usage{
get.results.table(
  object = NULL,
  which.analysis = NULL,
  results.table = NULL,
  which.assay = NULL,
  format = "df"
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.analysis}{A character specifying which analysis to retrieve results tables}

\item{results.table}{A data frame}

\item{which.assay}{A character specifying which assay to use.}

\item{format}{Output table format. One of:
\itemize{
\item df - Data.frame output. Preferred if user intends to do further data analysis
\item dt - Data.table output. Recommended for visualization and interactive table format. Data.table has interactive option to save table as csv or pdf, or copy contents into clipboard.
}}
}
\value{
list of table(s)
}
\description{
Retrieve result tables from Calibration Object.
}
\details{
Calibration Object (with specified analysis) or Data.frame table are accepted as inputs (Calibration Object takes precedent), and tables are reformatted and returned as data.frame or data.table, as specified by 'format' parameter.

If Calibration Object is provided, all result tables for specified analysis are retrieved.
}
\examples{
# get all results tables for precision analysis of uncalibrated data
results.tbl <-get.results.table(object = co,
                               which.analysis = "svp.analysis.uncalibrated",
                               format = 'dt') # 'dt' (datatable) or 'df' (dataframe)

# see which results.tables exist
names(results.tbl)

# print data.tables
print(results.tbl["replicate.statistics_results"])

print(results.tbl["rms.statistics_results"])
}
