% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_functions.R
\name{get.unique.feature.count}
\alias{get.unique.feature.count}
\title{Get Unique Feature Counts}
\usage{
get.unique.feature.count(object, verbose = TRUE, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{verbose}{print progress}

\item{which.assay}{specifies which assay to retrieve feature counts from (character)}
}
\value{
list
}
\description{
Get counts of variables for each unique feature present in specified assay of calibration object
}
