% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getAnalyses}
\alias{getAnalyses}
\title{Get Analyses}
\usage{
getAnalyses(object, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Specifies which assay to retrieve feature counts from (character)}
}
\value{
Character
}
\description{
Get names of available analyses for specified assay in Calibration Object.
}
\seealso{
\code{\link{svpAnalysis}}, \code{\link{mvpAnalysis}}
}
