% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getCalibrationInfo}
\alias{getCalibrationInfo}
\title{Get Calibration information}
\usage{
getCalibrationInfo(object, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character indicating which assay to retrive calibration information for. If unspecified, current Assay is reported.}
}
\value{
Calibration information is printed.
}
\description{
Reports calibration information for specified Assay
}
