% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getDatasets}
\alias{getDatasets}
\title{Get dataset names}
\usage{
getDatasets(object, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Specifies which assay to retrieve feature counts from (character)}
}
\value{
character
}
\description{
Get names of datasets stored in specified assay in calibration object
}
