% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getFeatures}
\alias{getFeatures}
\title{Get Features}
\usage{
getFeatures(object, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to retrieve features from.}
}
\value{
list of unique features
}
\description{
Get list of unique features from specified assay in calibration object
}
