% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getResults}
\alias{getResults}
\title{Get Results Table}
\usage{
getResults(
  object = NULL,
  which.results = NULL,
  which.data = NULL,
  results.table = NULL,
  which.assay = NULL,
  format = "df",
  max.page.length = 10
)
}
\arguments{
\item{object}{Calibration Object (input option 1)}

\item{which.results}{A character specifying function to retrieve results. Only specify for input option 1. One of: (
\itemize{
\item svp.analysis - returns replicate and rms statistics
\item mvp.analysis - results replicate and rms statistics
\item fit.calibration - returns calibration equations
}}

\item{results.table}{A data frame (input option 2)}

\item{which.assay}{A character specifying which assay to use. Only specify for input option 1.}

\item{format}{Output table format. One of:
\itemize{
\item df - Data frame. Preferred if user intends to do further data analysis
\item dt - Data table. Recommended for visualization and interactive table format. Data.table has interactive option to save table as csv or excel, or copy contents into clipboard.
}}

\item{max.page.length}{Default is 10. Maximal number of entries shown per page. Relevant only for datatable-formated tables (i.e., format = dt)}
}
\value{
list of table(s)
}
\description{
Returns table as data frame or interactive data table
}
\details{
Calibration Object (input option 1) or data frame (input option 2) are accepted as inputs (Calibration Object takes precedent), and tables are reformatted and returned as data.frame or data.table, as specified by 'format' parameter.
If Calibration Object is provided, all result tables for specified analysis are retrieved.
}
