% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getUniqueFeatureCount}
\alias{getUniqueFeatureCount}
\title{Get Unique Feature Counts}
\usage{
getUniqueFeatureCount(object, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Specifies which assay to retrieve feature counts from (character)}
}
\value{
data frame
}
\description{
Get unique feature countsfor each ufeature present in specified assay of calibration object
}
