% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{getUniqueFeatures}
\alias{getUniqueFeatures}
\title{Get Unique Features}
\usage{
getUniqueFeatures(df)
}
\arguments{
\item{df}{HR-pQCT dataset (data.frame). Following column names are expected: value, site, section, timePoint, phantom, parameter, scanDate}
}
\value{
list of unique.features (unique.features), number of variables per unique feature (N), and names of variables (variable)
}
\description{
Get unique features in input dataset
}
