% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invivo}
\alias{invivo}
\title{HR-pQCT in vivo patient reproducibility dataset}
\format{A data frame with columns:
\describe{
\item{scanID}{Scan identfier}
\item{site}{Site/scanner used to obtain meaures (e.g., Montreal, Toronto, etc.)}
\item{parameter}{Scanned parameter (e.g., Tb. vBMD, Ct. Ar., etc)}
\item{timePoint}{Time of scan; numeric (e.g., 0, 6, 12)}
\item{value}{Measured value; numeric}
\item{scanner}{HR-pQCT scanner (e.g., XCT, XCT2)}
}}
\source{
Shriners Hospital for Children, Mereo BioPharma
}
\usage{
invivo
}
\description{
HR-pQCT in vivo patient reproducibility dataset
}
\examples{
\dontrun{
# load data into global enviroment
 df <- invivo
}
}
\keyword{datasets}
