% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{invivo_unregistered}
\alias{invivo_unregistered}
\title{HR-pQCT in vivo patient reproducibility dataset (unregistered)}
\format{A data frame with columns:
\describe{
\item{scanID}{Scan identfier}
\item{site}{Site/scanner used to obtain meaures (e.g., Montreal, Toronto, etc.)}
\item{parameter}{Scanned parameter (e.g., Tb. vBMD, Ct. Ar., etc)}
\item{timePoint}{Time of scan; numeric (e.g., 0, 6, 12)}
\item{value}{Measured value; numeric}
\item{scanner}{HR-pQCT scanner (e.g., XCT, XCT2)}
\item{replicateSet}
\item{age}{Patient age (years); numeric}
\item{sex}{Patient sex; character (e.g., Male, Female)}
\item{weight}{Patient weight (kg); numeric}
\item{registration}{Registration status; character (e.g., CAS.registered, unregistered)}
}}
\source{
Shriners Hospital for Children, Mereo BioPharma
}
\usage{
invivo_unregistered
}
\description{
HR-pQCT in vivo patient reproducibility dataset (unregistered)
}
\examples{
\dontrun{
# load data into global enviroment
 df <- invivo_unregistered
}
}
\keyword{datasets}
