% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mean.var.plot}
\alias{mean.var.plot}
\title{mean-variance relationship}
\usage{
\method{mean}{var.plot}(
  object,
  which.assay = NULL,
  which.data = "uncalibrated",
  which.plot = "scatter",
  which.parameter = NULL,
  plot.outliers = F,
  return.plt.handle = F,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  point.size = 2,
  point.alpha = 1,
  error.squared = T
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}

\item{which.plot}{Character specifying which type of plot to generate. One of:
\itemize{
\item "scatter" -  Default. scatter plot of measure values vs. std and cv values
\item "box" - boxplot of std and cv values stratified by phantom section
}}

\item{which.parameter}{Character specfiying which parameters to plot. If unspecified, all parameters are plotted.}

\item{plot.outliers}{Logical specifying whether to include outliers (Default: false)}

\item{return.plt.handle}{Logical specifying whether to return plot handle. If TRUE, list of plot handles is returned. If FALSE, plots are printed and not returned.}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins}

\item{color.end}{Hue in [0,1] at which the viridis color map ends}

\item{point.size}{Numeric specifying size of scatter plot points. Only implemented if which.plot is "scatter". Default is 2.}

\item{point.alpha}{Numeric [0,1] specifying transparency of points. Only implemented if which.plot is "scatter". Default is 1.}

\item{error.squared}{Logical indicating whether to evaluate CV and STD (FALSE) or CV^2 and STD^2 (TRUE). Default is true.}
}
\value{
List of ggplot handles
}
\description{
Evalute relationship between replicate means and replicate std and cvs
}
