% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{mvp.analysis}
\alias{mvp.analysis}
\title{Multi variant precision/reproducibility analysis}
\usage{
mvp.analysis(
  object,
  which.data = "all",
  var2plot = "cv.value",
  show.data.table = TRUE,
  n.signif = 3,
  which.assay = NULL,
  trim.plot = FALSE,
  plot.flag = TRUE
)
}
\arguments{
\item{object}{calibration object}

\item{which.data}{specifies data to analysis (uncalibrated or calibrated)}

\item{show.data.table}{logical specifying whether to generate datatable summary of results}

\item{n.signif}{number of significant digits to report}

\item{which.assay}{specifies assay for analysis}

\item{trim.plot}{logical specifying whether to trim outliers from plots}

\item{plot.flag}{logical specifying whether to plot results}
}
\value{
calibration object
}
\description{
Longitudinal (i.e., multiple time points) and multicenter (i.e., multiple instruments) reproducibility/precision analysis of
repeated phantom/in vivo measurements. #' Precisions are calculated as root-mean-square standard deviation (rms.std) and
root-mean-square coefficient of variance (rms.cv). Results are saved in calibration object as 'mvp.analysis'.
}
