% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{mvpAnalysis}
\alias{mvpAnalysis}
\title{Multi variant precision/reproducibility analysis}
\usage{
mvpAnalysis(
  object,
  which.data = "all",
  which.phantom = NULL,
  replicate.strata = NULL,
  which.assay = NULL,
  n.signif = 3,
  verbose = T
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.data}{Character specifying which data to analyze. One of:
\itemize{
\item "all" (Default)
\item "uncalibrated"
\item "calibrated"
}}

\item{which.phantom}{Character vector specifying which phantom to include in multi-variant analysis.}

\item{replicate.strata}{Character vector specifying how a set of scan replicates are defined. Default replicate definition is stratification by site, timePoint, section, parameter, phantom. That is, all scans grouped by the specified features, and pooled within each strata prior to pooling across strata to compute parameter-specific rms-statistics. IMPORTANT. Do NOT use replicateSet feature as replicate.strata arguemnt, as it will no permit evaluation of multivariant precision.}

\item{which.assay}{Character specifying assay for analysis. If unspecified, set to current assay.}

\item{n.signif}{number of significant digits to report}

\item{verbose}{Logical specifying where to report progress.}
}
\value{
Calibration Object
}
\description{
Longitudinal (i.e., multiple time points) and multicenter (i.e., multiple instruments) precision analysis of
repeated phantom measurements. Precisions are calculated as root-mean-square standard deviation (rms.std) and
root-mean-square coefficient of variance (rms.cv). Results are saved in Calibration Object as 'mvp.analysis'.
}
\seealso{
\code{\link{svpAnalysis}}
}
