% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{mvpPlot}
\alias{mvpPlot}
\title{Multi-variant reproducibility boxplots}
\usage{
mvpPlot(
  object,
  which.data = "all",
  group.by = NULL,
  which.parameter = NULL,
  which.phantom = NULL,
  which.precision = "all",
  which.analysis = NULL,
  var2plot = "cv",
  which.assay = NULL,
  outliers = T,
  jitter.width = 0.1,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  point.size = 2,
  point.alpha = 1,
  box.alpha = 0.5,
  return.plt.handle = F,
  combine.plots = F,
  combine.ncols = 2,
  show.rms.statistic = F,
  show.legend.flag = T
)
}
\arguments{
\item{object}{Calibration Object.}

\item{which.data}{Character specifying which dataset to plot svp analysis for. One of:
\itemize{
\item "all" - Default
\item "uncalibrated"
\item "calibrated"
}}

\item{group.by}{Vector of features to stratify analysis by. If specified, rms-statistic is not overlaid on boxplot.}

\item{which.parameter}{Character specifying which parameters to plot. If unspecified, all parameters are plotted.}

\item{which.phantom}{Character specifying which phantoms to plot. If unspecified, all phantoms are plotted.}

\item{which.precision}{Character specifying which precision errors to plot. One of:
\itemize{
\item "all" - Default. Short and longitudinal, single- and multi-site precision errors
\item "short" - Short-term precision errors only
\item "long" - Longitudinal precision errors only
\item "single" - Single-site precision errors only
\item "multi" - Multi-site precision errors only
}}

\item{which.analysis}{Character indicating which mvp.analysis to get results from. If unspecified, available analyses will be used.}

\item{var2plot}{Character specifying which precision error metric to plot. One of:
\itemize{
\item cv - coefficient of variance
\item std - standard deviation
}}

\item{which.assay}{Character specifying which assay to plot.}

\item{outliers}{Logical specifying whether to omit outliers from analysis}

\item{jitter.width}{Numerical specifying width of jitter in boxplot. Default is 0.1.}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins. Default is 0.}

\item{color.end}{Hue in [0,1] at which the viridis color map ends. Default is 1.}

\item{point.size}{Numeric specifying size of scatter plot points.}

\item{point.alpha}{Numeric [0,1] specifying degree of transparency for points. Default is 1.}

\item{box.alpha}{Numeric [0,1] specifying degree of transparency for box plots. Default is 0.5.}

\item{return.plt.handle}{Logical indicating whether list of ggplot handle is returned. Default is false.}

\item{combine.plots}{Logical indicating whether plots are combined. Default is false.}

\item{combine.ncols}{Number of columns when combining plots (Only if combine.plot is true). Default is 2.}

\item{show.rms.statistic}{Logical indicating to overlay RMS statistic (rms-std or rms-cv) on boxplots. Default is false.}

\item{show.legend.flag}{Logical indicating to show legend}
}
\value{
list of ggplot objects (if return.plt.handle is true)
}
\description{
Generates boxplots summarizing results generated during mvp.analysis.
}
\seealso{
\code{\link{mvpAnalysis}}
}
