% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data_functions.R
\name{omit.outliers}
\alias{omit.outliers}
\title{Omit Outliers}
\usage{
omit.outliers(x, na.rm = TRUE, ...)
}
\arguments{
\item{x}{values}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
values
}
\description{
Credit: aL3xa, StackOVerflow (https://stackoverflow.com/a/4788102)
}
