% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Preprocess data}
\usage{
preprocess.data(object, analyze.which, new.assay.name, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{analyze.which}{list specifying which features to include in analysis (generate using analyzeWhich() function)}

\item{new.assay.name}{name of new preprocessed assay}

\item{which.assay}{specifies which assay to preprocess. Set to current.assay if which.assay is unspecified.}
}
\value{
calibration object
}
\description{
Preprocess HR-pQCT data for subsequent analysis
}
