% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_functions.R
\name{preprocessData}
\alias{preprocessData}
\title{Preprocess data}
\usage{
preprocessData(object, analyze.which, new.assay.name, which.assay = NULL)
}
\arguments{
\item{object}{Calibration Object}

\item{analyze.which}{List specifying which features to include in analysis (generate using analyzeWhich() function)}

\item{new.assay.name}{Name of new preprocessed assay}

\item{which.assay}{Character specifying which assay to preprocess. Set to default assay if which.assay is unspecified.}
}
\value{
Calibration Object
}
\description{
Preprocess HR-pQCT data for subsequent analysis
}
\seealso{
\code{\link{analyzeWhich}}
}
