% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{setDefaultAssay}
\alias{setDefaultAssay}
\title{Set Default Assay}
\usage{
setDefaultAssay(object, which.assay, verbose = TRUE)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies which assay is set to default (character)}

\item{verbose}{print progress}
}
\value{
calibration object with default assay set to specified assay
}
\description{
Set default (current) assay. Analyses are performed on default assay only.
}
