% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{svp.analysis}
\alias{svp.analysis}
\title{Single variant precision/reproducibility analysis}
\usage{
svp.analysis(
  object,
  which.data,
  stratify.by = "all",
  show.data.table = TRUE,
  n.signif = 3,
  which.assay = NULL,
  omit.outliers = FALSE
)
}
\arguments{
\item{object}{calibration object}

\item{which.data}{specifies data to analysis (uncalibrated or calibrated)}

\item{stratify.by}{features to stratify analysis by.}

\item{show.data.table}{logical specifying whether to generate datatable summary of results}

\item{n.signif}{number of significant digits to report}

\item{which.assay}{specifies assay for analysis}

\item{omit.outliers}{logical specifying whether to omit outliers from analysis}
}
\value{
calibration object
}
\description{
Short-term single-center reproducibility/precision analysis of repeated phantom/in vivo measurements.
Precisions are calculated as root-mean-square standard deviation (rms.std) and root-ean-square coefficient of variance (rms.cv).
Results are saved in calibration object as 'svp.analysis'.
}
