% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R, R/plot_functions.R
\name{svp.boxplot}
\alias{svp.boxplot}
\alias{fit.calibration}
\alias{svp.violinplot}
\title{Fit calibration curves}
\usage{
fit.calibration(
  object,
  reference.site = NULL,
  reference.time = "baseline",
  which.assay = NULL,
  n.signif = 3,
  show.data.table = TRUE,
  plot.flag = TRUE
)

svp.boxplot(
  object,
  var2plot = "cv.value",
  stratify.by = NULL,
  which.analysis = NULL,
  which.assay = NULL,
  show.outliers = FALSE,
  save.interactive.pooled.plot = FALSE,
  jitter.width = 0.1
)

svp.violinplot(
  object,
  var2plot = "cv.value",
  stratify.by = NULL,
  which.analysis = NULL,
  which.assay = NULL,
  show.outliers = FALSE,
  save.interactive.pooled.plot = FALSE,
  jitter.width = 0.1
)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{which assay to plot}

\item{var2plot}{precision error to plot (cv.value or std.value)}

\item{stratify.by}{features to stratify analysis by.}

\item{which.analysis}{which analysis to plot (character)}

\item{show.outliers}{logical specifying whether to omit outliers from analysis}

\item{jitter.width}{width of jitter in boxplot}
}
\value{
calibration object

calibration object

calibration object
}
\description{
Fit pairwise and

Generates boxplot summarizing results from svp.analysis.

Generates violin plot summarizing results from svp.analysis.
}
