% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{svp.plot}
\alias{svp.plot}
\alias{mvp.plot}
\title{Short-term reproducibility boxplot}
\usage{
svp.plot(
  object,
  which.data = "uncalibrated",
  group.by = NULL,
  var2plot = "cv",
  which.assay = NULL,
  omit.outliers = FALSE,
  jitter.width = 0.1
)

mvp.plot(
  object,
  which.data = "uncalibrated",
  group.by = NULL,
  var2plot = "cv",
  which.assay = NULL,
  omit.outliers = FALSE,
  jitter.width = 0.1
)
}
\arguments{
\item{object}{Calibration Object.}

\item{which.data}{Character specifying which dataset to plot svp analysis for. One of:
\itemize{
\item uncalibrated - Default
\item calibrated
}}

\item{group.by}{Vector of features to stratify analysis by. If specified, rms-statistic is not overlaid on boxplot.}

\item{var2plot}{Character specifying which precision error metric to plot. One of:
\itemize{
\item cv - coefficient of variance
\item std - standard deviation
}}

\item{which.assay}{Character specifying which assay to plot.}

\item{omit.outliers}{Logical specifying whether to omit outliers from analysis}

\item{jitter.width}{Numerical specifying width of jitter in boxplot. Default = 0.1.}
}
\value{
ggplot object

ggplot object
}
\description{
Generates boxplot summarizing results generated during svp.analysis.

Generates boxplot summarizing results generated during svp.analysis.
}
\seealso{
svp.analysis

svp.analysis
}
