% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{svpAnalysis}
\alias{svpAnalysis}
\title{Single variant precision/reproducibility analysis}
\usage{
svpAnalysis(
  object,
  which.data = "uncalibrated",
  replicate.strata = NULL,
  n.signif = 3,
  which.assay = NULL,
  verbose = T
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.data}{A character specifying which data to analyze. One of:
\itemize{
\item uncalibrated (Default)
\item calibrated
}}

\item{replicate.strata}{Vector specifying how a set of scan replicates are defined. If specified, overrides replicateSet definition. If replicateSet definition not present in Calibration Object and replicate.strata is not specified, default replicate definition is stratification by site, timePoint, section, parameter, phantom. That is, all scans grouped by the specified features, and pooled within each strata prior to pooling across strata to compute parameter-specific rms-statistics.}

\item{n.signif}{Number of significant digits reported.}

\item{which.assay}{A character specifying assay for analysis. If unspecified, set to current assay.}

\item{verbose}{Logical specifying where to report progress.}
}
\value{
Calibration Object
}
\description{
Short-term single-center precision analysis of repeated phantom/in vivo measurements.
Parameter-specific precision errors are calculated as root-mean-square standard deviation (rms.std) and root-mean-square coefficient of variance (rms.cv).
Results are saved in calibration object as 'svp.analysis'.
}
\seealso{
\code{\link{mvpAnalysis}}, \code{\link{defineReplicateSet}}
}
