% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{svpPlot}
\alias{svpPlot}
\title{Short-term reproducibility boxplot}
\usage{
svpPlot(
  object,
  which.data = "uncalibrated",
  group.by = NULL,
  which.parameter = NULL,
  which.phantom = NULL,
  var2plot = "cv",
  which.assay = NULL,
  outliers = TRUE,
  jitter.width = 0.1,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  point.size = 2,
  point.alpha = 1,
  box.alpha = 0.5,
  txt.size = NULL,
  xlab.direction = NULL,
  xlab.hjust = NULL,
  xlab.vjust = NULL,
  show.p = T,
  color.option.v2 = NULL,
  color.begin.v2 = NULL,
  color.end.v2 = NULL,
  ylim = NULL
)
}
\arguments{
\item{object}{Calibration Object.}

\item{which.data}{Character specifying which dataset to plot svp analysis for. One of:
\itemize{
\item uncalibrated - Default
\item calibrated
}}

\item{group.by}{Vector of features to stratify analysis by. If specified, rms-statistic is not overlaid on boxplot.}

\item{which.parameter}{Character specifying which parameters to plot. If unspecified, all parameters are plotted.}

\item{which.phantom}{Character specifying which phantoms to plot. If unspecified, all available phantoms are plotted.}

\item{var2plot}{Character specifying which precision error metric to plot. One of:
\itemize{
\item cv - coefficient of variance
\item std - standard deviation
}}

\item{which.assay}{Character specifying which assay to plot.}

\item{outliers}{Logical specifying whether to omit outliers from analysis}

\item{jitter.width}{Numerical specifying width of jitter in boxplot. Default is 0.1.}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins. Default is 0.}

\item{color.end}{Hue in [0,1] at which the viridis color map ends. Default is 1.}

\item{point.size}{Numeric specifying size of scatter plot points. Default is 2.}

\item{point.alpha}{Numeric [0,1] specifying degree of transparency for points. Default is 1.}

\item{box.alpha}{Numeric [0,1] specifying degree of transparency for box plots. Default is 0.5.}

\item{txt.size}{Numeric specifying size of text}

\item{xlab.direction}{Numeric specifying direciton (i.e., angle) of x axis labels}

\item{xlab.hjust}{Horizontal justification of x axis labels; 0 is left-justified, 0.5 is centered, and 1 is right-justified.}

\item{xlab.vjust}{Vertical justification of x axis labels;}

\item{show.p}{Logical specifying if p-values for Kruskal-Wallis rank sum test are shown. Default is true.}

\item{color.option.v2}{Same as color.option, provides additional flexibility to specify color pallete for second grouping variable if length(group.by) == 2. If unspecified, set to color.option.}

\item{color.begin.v2}{Same as color.begin, provides additional flexibility to specify hue for second grouping variable if length(group.by) == 2. If unspecified, set to color.begin}

\item{color.end.v2}{Same as color.end, provides additional flexibility to specify hue for second grouping variable if length(group.by) == 2. If unspecified, set to color.end}

\item{ylim}{Y axis limits. If unspecified, then set automatically.}
}
\value{
ggplot object
}
\description{
Generates boxplot summarizing results generated during svp.analysis. Additionally, Kruskal-Wallis rank sum test is performed (using group.by variable for data stratification) and p-values are shown for each comparison group.
}
\seealso{
\code{\link{svpAnalysis}}
}
