% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{deleteAssay}
\alias{deleteAssay}
\title{Delete Assay}
\usage{
deleteAssay(object, which.assay, verbose = TRUE)
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{A character specifying which assay to delete.}

\item{verbose}{Default is TRUE. Logical specifying whether to print progress.}
}
\value{
calibration object with specified assay deleted.
}
\description{
Delete assay from calibration object
}
\details{
If specified assay was current assay, new current assay is assigned.
}
