% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_set_functions.R
\name{get.analyses}
\alias{get.analyses}
\title{Get Analyses}
\usage{
get.analyses(object, verbose = TRUE, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{verbose}{print progress}

\item{which.assay}{specifies which assay to retrieve feature counts from (character)}
}
\value{
character
}
\description{
Get names of avaialble analyses for specified assay in calibration object
}
