% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.assay}
\alias{get.assay}
\title{Get Assay}
\usage{
get.assay(object, which.assays = "default")
}
\arguments{
\item{object}{calibration object}

\item{which.assays}{which assay(s) to return. Accepted arguments are "default" or "all"}
}
\value{
name(s) of assay(s); character
}
\description{
get name of assays (current or all)
}
