% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.assay}
\alias{get.assay}
\title{Get Assay}
\usage{
get.assay(object, which.assays = "default")
}
\arguments{
\item{object}{Calibration Object}

\item{which.assays}{A character specifying which assay(s) to return. Accepted arguments are "default" or "all"
\itemize{
\item default - current assay
\item all - all existing assays
}}
}
\value{
name(s) of assay(s); character
}
\description{
Get name of assays
}
