% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.unique.feature.count}
\alias{get.unique.feature.count}
\title{Get Unique Feature Counts}
\usage{
get.unique.feature.count(object, which.assay = NULL)
}
\arguments{
\item{object}{calibration object}

\item{which.assay}{specifies which assay to retrieve feature counts from (character)}
}
\value{
data.frame
}
\description{
Get unique feature countsfor each ufeature present in specified assay of calibration object
}
