% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_functions.R
\name{identifyReference}
\alias{identifyReference}
\title{Identify Reference Site}
\usage{
identifyReference(object, which.assay = NULL, which.data = "uncalibrated")
}
\arguments{
\item{object}{Calibration Object}

\item{which.assay}{Character specifying which assay to identify reference site for.}

\item{which.data}{Character specifying which data to identify reference site for.}
}
\value{
Character
}
\description{
Sites are ranked according to mean-squared error calculated between site-specific parameters and overall median parameter values (pooled across all timePoints and sites), and top ranking site is returned.
}
\seealso{
\code{\link{consistencyAnalysis}}, \code{\link{consistencyPlot}}
}
