% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{svp.analysis}
\alias{svp.analysis}
\title{Single variant precision/reproducibility analysis}
\usage{
svp.analysis(
  object,
  which.data,
  stratify.by = NULL,
  n.signif = 3,
  which.assay = NULL,
  omit.outliers = FALSE
)
}
\arguments{
\item{object}{Calibration Object}

\item{which.data}{A character that specifies which data to analysis. One of  (uncalibrated or calibrated)
\itemize{
\item uncalibrated
\item calibrated
}}

\item{stratify.by}{features to stratify analysis by. If left unspecified, default grouping is used (recommended), i.e., site, timePoint, section, parameter, phantom}

\item{n.signif}{number of significant digits to report}

\item{which.assay}{specifies assay for analysis}

\item{omit.outliers}{logical specifying whether to omit outliers from analysis. If TRUE, outliers will be saved as separate dataset within current Assay.}
}
\value{
calibration object
}
\description{
Short-term single-center reproducibility/precision analysis of repeated phantom/in vivo measurements.
Precisions are calculated as root-mean-square standard deviation (rms.std) and root-ean-square coefficient of variance (rms.cv).
Results are saved in calibration object as 'svp.analysis'.
}
