% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{svpPlot}
\alias{svpPlot}
\title{Short-term reproducibility boxplot}
\usage{
svpPlot(
  object,
  which.data = "uncalibrated",
  group.by = NULL,
  which.parameter = NULL,
  var2plot = "cv",
  which.assay = NULL,
  outliers = TRUE,
  jitter.width = 0.1,
  color.option = "cividis",
  color.begin = 0,
  color.end = 1,
  point.size = 2,
  point.alpha = 1,
  box.alpha = 0.5
)
}
\arguments{
\item{object}{Calibration Object.}

\item{which.data}{Character specifying which dataset to plot svp analysis for. One of:
\itemize{
\item uncalibrated - Default
\item calibrated
}}

\item{group.by}{Vector of features to stratify analysis by. If specified, rms-statistic is not overlaid on boxplot.}

\item{which.parameter}{Character specifying which parameters to plot. If unspecified, all parameters are plotted.}

\item{var2plot}{Character specifying which precision error metric to plot. One of:
\itemize{
\item cv - coefficient of variance
\item std - standard deviation
}}

\item{which.assay}{Character specifying which assay to plot.}

\item{outliers}{Logical specifying whether to omit outliers from analysis}

\item{jitter.width}{Numerical specifying width of jitter in boxplot. Default is 0.1.}

\item{color.option}{Character indicating which color map option to use (from viridis palette). One of:
\itemize{
\item "magma" (or "A")
\item "inferno" (or "B")
\item "plasma" (or "C")
\item "viridis" (or "D")
\item "cividis" (or "E") - Default
}}

\item{color.begin}{Hue in [0,1] at which the viridis color map begins. Default is 0.}

\item{color.end}{Hue in [0,1] at which the viridis color map ends. Default is 1.}

\item{point.size}{Numeric specifying size of scatter plot points. Default is 2.}

\item{point.alpha}{Numeric [0,1] specifying degree of transparency for points. Default is 1.}

\item{box.alpha}{Numeric [0,1] specifying degree of transparency for box plots. Default is 0.5.}
}
\value{
ggplot object
}
\description{
Generates boxplot summarizing results generated during svp.analysis. Additionally, Kruskal-Wallis rank sum test is performed (using group.by variable for data stratification) and p-values are shown for each comparison group.
}
\seealso{
\code{\link{svpAnalysis}}
}
